Algorithm now uses hash(salt + secret)
This makes it easier for implementations to store that hash in memory
reducing viability of scraping memory for users password.  The
implementations can use this hash to generate passwords for any number
of sites.

NOTE: Access to this hash will still compromise all generated passwords,
but the users master password should still be safe.  If hash is
compromised a new master password should be used to regenerate all
passwords.